split code
