+++ mcp/lib/component-meta.js
@@ -1,0 +1,193 @@
+// [name]: [description (1 sentence, function only), tags]
+export const COMPONENT_META = {
+  // ── DataDisplay
+  Table:                      ['Display tabular data with sorting, selection, and custom cell slots.', ['table', 'data', 'list']],
+  TableHeader:                ['Header cell for Table, supports text or checkbox variant.', ['table', 'data']],
+  TableCell:                  ['Body cell for Table, supports content/chip/datetime/icon/image variants.', ['table', 'data']],
+  TableCellContent:           ['Two-line cell content with title and description.', ['table', 'data']],
+  TableLink:                  ['Wraps table cell content as a link.', ['table', 'data']],
+  ExpandingTable:             ['Table with collapsible/expandable nested rows.', ['table', 'data', 'tree']],
+  ExpandingTableRow:          ['Single row inside ExpandingTable.', ['table', 'data']],
+  ExpandingHeaderRow:         ['Header row for ExpandingTable.', ['table', 'data']],
+  Pagination:                 ['Page navigation with prev/next and numbered buttons.', ['pagination', 'navigation']],
+  PrevNext:                   ['Previous or next button used inside Pagination.', ['pagination', 'navigation']],
+  PaginationNumber:           ['Numbered page button used inside Pagination.', ['pagination', 'navigation']],
+  Overview:                   ['Group of metric cards showing key stats with icons.', ['stats', 'dashboard', 'metrics']],
+  OverviewIcon:               ['Colored icon badge inside OverviewItem.', ['icon', 'stats']],
+  OverviewItem:               ['Single metric card with title, value, and trend indicator.', ['stats', 'dashboard', 'metrics']],
+  CustomerCard:               ['Display customer or beneficiary details.', ['customer', 'card', 'profile']],
+  TransferSummary:            ['Show money transfer summary with amount, fee, and FX rate.', ['transfer', 'finance', 'summary']],
+  ListItem:                   ['Versatile list row for timeline, webhook, payment, and accordion layouts.', ['list', 'timeline', 'item']],
+  ListUrl:                    ['Display a labelled URL with copy button.', ['list', 'url', 'copy']],
+  ListDetail:                 ['Label-value row in horizontal or vertical layout.', ['list', 'detail', 'label']],
+  InfoCard:                   ['Content card with title, loading state, and header/footer slots.', ['card', 'content']],
+  ProgressBar:                ['Step-based linear progress bar.', ['progress', 'steps']],
+  FloatContent:               ['Floating side panel with title, description, and actions.', ['panel', 'overlay', 'sidebar']],
+  HorizontalProgressbar:      ['Horizontal bar showing current vs max value.', ['progress', 'bar']],
+  OcOnboardingListItem:       ['Onboarding checklist row with number badge and action button.', ['onboarding', 'checklist']],
+  OcOnboardingRecap:          ['Collapsible recap card for onboarding sections.', ['onboarding', 'recap', 'summary']],
+  OcAccountVerificationContent: ['Container for account verification steps.', ['verification', 'onboarding']],
+  OcAccountVerificationBox:   ['Selectable verification option card.', ['verification', 'onboarding', 'option']],
+  NewTable:                   ['Table with row expansion, selection, and named column slots.', ['table', 'data', 'list']],
+  OcRowItemData:              ['List of labelled data items with optional chip and copy button.', ['list', 'data', 'detail']],
+  // ── Disclosure
+  Accordion:                  ['Collapsible section with header and body.', ['accordion', 'collapse', 'disclosure']],
+  OnboardingAccordion:        ['Accordion with onboarding status variant.', ['accordion', 'onboarding']],
+  OnboardingProductOption:    ['Selectable product option card with image.', ['onboarding', 'option', 'card']],
+  Tabs:                       ['Tab bar for switching between views.', ['tabs', 'navigation']],
+  PaymentTab:                 ['Tab selector for payment methods.', ['tabs', 'payment']],
+  Steps:                      ['Step indicator for multi-step flows.', ['steps', 'wizard', 'progress']],
+  Step:                       ['Single step inside Steps.', ['steps', 'wizard']],
+  Variants:                   ['Group of selectable option cards.', ['variants', 'option', 'select']],
+  Variant:                    ['Single selectable option card with icon and label.', ['variants', 'option']],
+  SwitchTabs:                 ['Segmented control for switching between a few options.', ['tabs', 'toggle', 'switch']],
+  // ── Layout / Elements
+  Header:                     ['Top-level application header shell.', ['header', 'layout', 'navigation']],
+  SubHeader:                  ['Secondary header below the main header.', ['header', 'layout']],
+  HeaderLeft:                 ['Left slot of the Header.', ['header', 'layout']],
+  HeaderCenter:               ['Center slot of the Header with back button support.', ['header', 'layout']],
+  HeaderRight:                ['Right slot of the Header with action buttons.', ['header', 'layout']],
+  TabToSelect:                ['Tab bar inside Header for route switching.', ['tabs', 'header', 'navigation']],
+  Checklist:                  ['Checklist step item with index, label, and completion state.', ['checklist', 'onboarding']],
+  PageTitle:                  ['Page title bar with back button and primary/secondary actions.', ['page', 'title', 'header']],
+  PrimaryActions:             ['Primary action buttons or dropdown inside PageTitle.', ['button', 'actions']],
+  PageTitleRight:             ['Right action area of PageTitle.', ['page', 'title', 'actions']],
+  Sidebar:                    ['Main application sidebar with expandable menus.', ['sidebar', 'navigation', 'layout']],
+  SidebarSubmenu:             ['Expandable submenu inside Sidebar.', ['sidebar', 'navigation']],
+  AccountSetup:               ['Account setup progress widget inside Sidebar.', ['sidebar', 'onboarding', 'progress']],
+  SidebarHead:                ['Header section of a Sidebar menu group.', ['sidebar', 'navigation']],
+  SideBarMenu:                ['Single menu item in Sidebar.', ['sidebar', 'navigation', 'menu']],
+  SidebarSubMenuItem:         ['Child menu item inside SidebarSubmenu.', ['sidebar', 'navigation']],
+  SidebarFooter:              ['Footer section of the Sidebar.', ['sidebar', 'layout']],
+  SidebarFeatureBanners:      ['Feature promotion banners inside Sidebar.', ['sidebar', 'banner', 'promo']],
+  AccountSetupProgress:       ['Circular progress indicator for account setup in Sidebar.', ['sidebar', 'progress', 'onboarding']],
+  SubSidebar:                 ['Secondary sidebar for sub-navigation.', ['sidebar', 'navigation']],
+  EmptyPage:                  ['Placeholder shown when a section has no data.', ['empty', 'placeholder', 'state']],
+  AdditionalContent:          ['Side panel with customer info, overview, and action buttons.', ['panel', 'sidebar', 'content']],
+  InfoCardButton:             ['Selectable card button with title and content.', ['card', 'button', 'select']],
+  Skeleton:                   ['Loading placeholder that mimics content shape.', ['loading', 'skeleton', 'placeholder']],
+  FeatureOverviewCard:        ['Card listing product features with icons.', ['card', 'features', 'overview']],
+  SandboxBanner:              ['Top banner indicating sandbox/test mode.', ['banner', 'sandbox', 'alert']],
+  TopBanner:                  ['Dismissible top-of-page announcement banner.', ['banner', 'alert', 'notification']],
+  Box:                        ['Clickable card with icon, title, and description.', ['card', 'box', 'button']],
+  OcSection:                  ['Content section wrapper with optional title.', ['section', 'layout']],
+  AuthenticationOption:       ['Selectable authentication method card.', ['auth', 'option', 'card']],
+  VerificationOption:         ['Selectable verification method card.', ['verification', 'option', 'card']],
+  // ── Feedback
+  Chip:                       ['Status or category badge with color and style variants.', ['chip', 'badge', 'tag', 'status']],
+  Snackbar:                   ['Toast notification with icon and action buttons.', ['toast', 'notification', 'alert']],
+  Banner:                     ['Inline notification banner with dismiss support.', ['banner', 'notification', 'alert']],
+  ShareIcon:                  ['Icon button for sharing via a channel (e.g. WhatsApp).', ['share', 'icon', 'social']],
+  Tag:                        ['Small colored label for status annotation.', ['tag', 'badge', 'label', 'status']],
+  // ── Form
+  BaseInput:                  ['Form field wrapper with label, hint, and error slot.', ['form', 'input', 'field']],
+  Button:                     ['Action button with primary/secondary/destructive variants.', ['button', 'action', 'form']],
+  CardInput:                  ['Credit card number input with card type detection.', ['input', 'payment', 'card', 'form']],
+  Checkbox:                   ['Single checkbox input.', ['checkbox', 'form', 'input']],
+  CheckboxesGroup:            ['Group of checkboxes with select-all support.', ['checkbox', 'form', 'group']],
+  Criteria:                   ['Read-only pass/fail criteria indicator.', ['criteria', 'validation', 'form']],
+  Input:                      ['Text input with icon, prefix, and password toggle.', ['input', 'text', 'form']],
+  InputOption:                ['Tag-style input for adding multiple free-text values.', ['input', 'tags', 'form']],
+  PhoneInput:                 ['Phone number input with country code selector.', ['input', 'phone', 'form']],
+  Radio:                      ['Single radio button input.', ['radio', 'form', 'input']],
+  RadioGroup:                 ['Group of radio buttons.', ['radio', 'form', 'group']],
+  RangeInput:                 ['Dual numeric input for a min/max range.', ['input', 'range', 'form']],
+  Select:                     ['Dropdown with single/multi select, filtering, and async search.', ['select', 'dropdown', 'form', 'filter']],
+  Option:                     ['Single option item inside Select.', ['select', 'option', 'form']],
+  GroupOptions:               ['Labeled group of options inside Select.', ['select', 'group', 'form']],
+  Toggle:                     ['On/off toggle switch.', ['toggle', 'switch', 'form']],
+  LinkInput:                  ['Input for social or custom links with platform selector.', ['input', 'link', 'social', 'form']],
+  TextArea:                   ['Multi-line text input with auto-resize.', ['textarea', 'input', 'form']],
+  Calendar:                   ['Standalone date or date-range calendar picker.', ['calendar', 'date', 'picker', 'form']],
+  DatePicker:                 ['Date or range picker input with popover calendar.', ['datepicker', 'date', 'form']],
+  Slider:                     ['Range or single value slider.', ['slider', 'range', 'form']],
+  SectionItem:                ['Settings row with title, toggle, and action slot.', ['settings', 'form', 'toggle']],
+  SelectOptions:              ['Visual option picker in grid or list layout.', ['select', 'options', 'form', 'grid']],
+  SingleFileUpload:           ['Single file upload with preview and size limit.', ['upload', 'file', 'form']],
+  MultipleUploadFile:         ['Multiple file upload with preview grid.', ['upload', 'file', 'form', 'gallery']],
+  NumberInput:                ['Numeric input with increment/decrement buttons.', ['input', 'number', 'form']],
+  ComplexCalendar:            ['Full calendar panel with range selection and shortcuts.', ['calendar', 'date', 'range', 'form']],
+  ComplexDatePicker:          ['Date range picker input with label and format options.', ['datepicker', 'date', 'range', 'form']],
+  ModalCropper:               ['Modal-based image cropper.', ['upload', 'image', 'crop', 'modal']],
+  Cropper:                    ['Inline image cropper.', ['upload', 'image', 'crop']],
+  EmojiPicker:                ['Emoji selector popover input.', ['emoji', 'input', 'form']],
+  TimePopup:                  ['Time picker popup.', ['time', 'picker', 'form']],
+  TimePicker:                 ['Time picker input with popover clock.', ['time', 'picker', 'form']],
+  // ── Media
+  Avatar:                     ['User or business avatar in multiple sizes.', ['avatar', 'image', 'user']],
+  Icon:                       ['SVG icon by name from the OrchidUI icon set.', ['icon', 'svg', 'media']],
+  // ── Overlay
+  Dropdown:                   ['Popover dropdown menu attached to a trigger.', ['dropdown', 'menu', 'overlay']],
+  DropdownItem:               ['Single item row inside a Dropdown.', ['dropdown', 'menu', 'item']],
+  Tooltip:                    ['Hover or click tooltip/popover.', ['tooltip', 'popover', 'overlay']],
+  Modal:                      ['Dialog modal with header, body, and footer.', ['modal', 'dialog', 'overlay']],
+  OcModalPage:                ['Full-screen modal page with back button.', ['modal', 'page', 'overlay']],
+  ConfirmationModal:          ['Pre-styled confirm/delete/warning dialog.', ['modal', 'confirm', 'dialog', 'overlay']],
+  Popper:                     ['Low-level Popper.js positioning wrapper.', ['overlay', 'popper', 'positioning']],
+  SupportMenu:                ['Floating help/support menu with grouped links.', ['support', 'help', 'menu', 'overlay']],
+  CopyTooltip:                ['Wrap any element to copy its value to clipboard.', ['copy', 'clipboard', 'tooltip']],
+  WhiteList:                  ['Feature gate card shown when plan does not include a feature.', ['upgrade', 'gate', 'plan']],
+  Theme:                      ['Wrap app to apply light/dark color mode.', ['theme', 'dark-mode', 'layout']],
+  // ── FormBuilder
+  FormBuilder:                ['Dynamic form renderer from a JSON schema.', ['form', 'builder', 'dynamic', 'schema']],
+  // ── DataTable
+  DataTable:                  ['Data table with filtering, pagination, search, and bulk actions.', ['table', 'data', 'filter', 'pagination', 'crud']],
+  FilterSearch:               ['Search input inside DataTable filter toolbar.', ['filter', 'search', 'table']],
+  FilterSearchFor:            ['Active filter chips with remove controls.', ['filter', 'search', 'table']],
+  FilterForm:                 ['Filter form panel driven by JSON schema, used inside DataTable.', ['filter', 'form', 'table']],
+  // ── Elements
+  Draggable:                  ['Drag-and-drop wrapper for reorderable lists.', ['drag', 'drop', 'sort', 'list']],
+  DraggableList:              ['Styled draggable list with icon and action slots.', ['drag', 'drop', 'sort', 'list']],
+  // ── StoreDesign
+  ThumbnailTheme:             ['Theme selection card with preview and activate/publish actions.', ['theme', 'store', 'design']],
+  ThumbnailSection:           ['Section template card with preview and add action.', ['section', 'store', 'design']],
+  SubPlanCard:                ['Subscription plan tier card listing features.', ['plan', 'subscription', 'pricing']],
+  // ── Dashboard
+  LineChart:                  ['Time-series line chart for trend visualization.', ['chart', 'line', 'analytics', 'dashboard']],
+  PieChart:                   ['Pie or donut chart for proportional data.', ['chart', 'pie', 'analytics', 'dashboard']],
+  BarChart:                   ['Bar chart for comparing values across categories.', ['chart', 'bar', 'analytics', 'dashboard']],
+  BarRaceChart:               ['Animated bar race chart for changing rankings.', ['chart', 'animated', 'analytics', 'dashboard']],
+  Card:                       ['Dashboard card container with title and content.', ['card', 'dashboard']],
+  OcOverviewPieChart:         ['Pie chart for overview/summary widgets.', ['chart', 'pie', 'overview', 'dashboard']],
+  OverviewMonthlyBarChart:    ['Monthly bar chart for overview reporting.', ['chart', 'bar', 'overview', 'dashboard']],
+  ColorPicker:                ['Color picker input with hex/rgb support.', ['color', 'picker', 'input']],
+  TextEditor:                 ['Rich text (WYSIWYG) editor.', ['editor', 'richtext', 'wysiwyg', 'input']],
+  ColorPickerPopup:           ['Popover color picker for inline color selection.', ['color', 'picker', 'popover']],
+  OnboardingProgressbar:      ['Progress bar for onboarding completion.', ['progress', 'onboarding', 'dashboard']],
+  CodeBlock:                  ['Syntax-highlighted code block.', ['code', 'syntax', 'display']],
+  LottieAnimation:            ['Renders a Lottie JSON animation.', ['animation', 'lottie', 'media']],
+}
+
+export const DESCRIPTIONS = Object.fromEntries(Object.entries(COMPONENT_META).map(([k, [d]]) => [k, d]))
+export const TAGS         = Object.fromEntries(Object.entries(COMPONENT_META).map(([k, [, t]]) => [k, t]))
+
+export const CORE_COMPONENT_NAMES = new Set([
+  'Table','TableHeader','TableCell','TableCellContent','TableLink','ExpandingTable',
+  'ExpandingTableRow','ExpandingHeaderRow','Pagination','PrevNext','PaginationNumber',
+  'Overview','OverviewIcon','OverviewItem','CustomerCard','TransferSummary','ListItem',
+  'ListUrl','ListDetail','InfoCard','ProgressBar','FloatContent','HorizontalProgressbar',
+  'OcOnboardingListItem','OcOnboardingRecap','OcAccountVerificationContent',
+  'OcAccountVerificationBox','NewTable','OcRowItemData','Accordion','OnboardingAccordion',
+  'OnboardingProductOption','Tabs','PaymentTab','Steps','Step','Variants','Variant',
+  'SwitchTabs','Header','SubHeader','HeaderLeft','HeaderCenter','HeaderRight','TabToSelect',
+  'Checklist','PageTitle','PrimaryActions','PageTitleRight','Sidebar','SidebarSubmenu',
+  'AccountSetup','SidebarHead','SideBarMenu','SidebarSubMenuItem','SidebarFooter',
+  'SidebarFeatureBanners','AccountSetupProgress','SubSidebar','EmptyPage','AdditionalContent',
+  'InfoCardButton','Skeleton','FeatureOverviewCard','SandboxBanner','TopBanner','Box',
+  'OcSection','AuthenticationOption','VerificationOption','Chip','Snackbar','Banner',
+  'ShareIcon','Tag','BaseInput','Button','CardInput','Checkbox','CheckboxesGroup','Criteria',
+  'Input','InputOption','PhoneInput','Radio','RadioGroup','RangeInput','Select','Option',
+  'GroupOptions','Toggle','LinkInput','TextArea','Calendar','DatePicker','Slider',
+  'SectionItem','SelectOptions','SingleFileUpload','MultipleUploadFile','NumberInput',
+  'Avatar','Icon','Dropdown','DropdownItem','Tooltip','Modal','OcModalPage',
+  'ConfirmationModal','Popper','SupportMenu','CopyTooltip','WhiteList','Theme','FormBuilder',
+  'DataTable','FilterSearch','FilterSearchFor','FilterForm','ComplexCalendar',
+  'ComplexDatePicker','ModalCropper','Cropper','Draggable','DraggableList','EmojiPicker',
+  'ThumbnailTheme','ThumbnailSection','SubPlanCard','TimePopup','TimePicker',
+])
+
+export const DASHBOARD_COMPONENT_NAMES = new Set([
+  'LineChart','PieChart','BarChart','BarRaceChart','Card','OcOverviewPieChart',
+  'OverviewMonthlyBarChart','ColorPicker','TextEditor','ColorPickerPopup',
+  'OnboardingProgressbar','CodeBlock','LottieAnimation',
+])
